backend/app/schemas/job.py
import uuid
from datetime import datetime

from pydantic import BaseModel


class JobResponse(BaseModel):
    id: uuid.UUID
    url_id: uuid.UUID
    viewport_label: str
    viewport_width: int
    viewport_height: int
    status: str
    created_at: datetime
    started_at: datetime | None
    completed_at: datetime | None
    error_message: str | None

    model_config = {"from_attributes": True}
